Continued realisation of data layer

+++ app/src/main/java/com/ponkratov/weatherapp/data/model/CityDTO.kt
@@ -3,12 +3,12 @@ package com.ponkratov.weatherapp.data.model
 import com.google.gson.annotations.SerializedName
 
 data class CityDTO(
-    @SerializedName("id") var id: Int? = null,
+    @SerializedName("id") var id: Long? = null,
     @SerializedName("name") var name: String? = null,
     @SerializedName("country_code") var countryCode: String? = null,
     @SerializedName("country") var country: String? = null,
     @SerializedName("latitude") var latitude: Double? = null,
     @SerializedName("longitude") var longitude: Double? = null,
-    @SerializedName("elevation") var elevation: Int? = null,
+    @SerializedName("elevation") var elevation: Long? = null,
     @SerializedName("timezone") var timezone: String? = null
 )
